% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_box}
\alias{plot_box}
\title{Boxplot graphic}
\usage{
plot_box(data, x, y, z, ylab = "", xlab = "", lgl = "", lgd = "right",
  brk = NULL, font = 1)
}
\arguments{
\item{data}{raw data}

\item{x}{Axis x variable}

\item{y}{Axis y variable}

\item{z}{Group variable}

\item{ylab}{Title for the axis y}

\item{xlab}{Title for the axis x}

\item{lgl}{Title for the legend}

\item{lgd}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{font}{letter size in plot}
}
\value{
boxplot
}
\description{
Function use the raw data for made a boxplot graphic
}

