% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Principal component analisys graphic}
\usage{
plot_PCA(data, habillage, quali.sup, quanti.sup)
}
\arguments{
\item{data}{matrix with numeric data}
}
\value{
PCA biplot graph
}
\description{
Function use to plot biplot principal component analisys
}

